requirements.txt
requests==2.24.0
beautifulsoup4==4.9.1
pywebview==3.3.4
click==7.1.2